Parralax working!!
Translate z, translate y, and scale have to all be declared at the same time in the same line, but then if I want to make the parallax go slower I just need to push it farther backwards in space, scale it up more, and make it start higher up on the screen.

Not sure yet how to make it set back to 0 at the bottom, or if it will mess with stuff after it yet.