Fixed sticky rail and adjusted main stage settings to work cleaner
The buffer and trigger boxes were inside of the writing-rail thats why sticky didn't seem to work before.
I also made it so instead of the whole page being a flex box that centered the main stage, the main stage was set absolutely in the page.
(I also removed some pointless max-width and aspect ratio rules that don't have any effect when I have explicitly set the size for the main-stage)