Fixed a couple bugs with the animation framework
The page will only start triggering adding or removing classes as soon as the person starts to actually scroll. Before it was firing as soon as the page loaded for some reason.

I also somehow had the data-remove-visible class running backwards so it was pretty much the same as the add logic. Now the add and remove logic work as they should.